Rename: 패키지 정리, 공통 파일들 common 패키지로 옮김
