+++ app/src/main/AndroidManifest.xml
@@ -118,7 +118,7 @@
                 android:value="" />
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name=".common.MainActivity"
             android:exported="true">
             <meta-data
                 android:name="android.app.lib_name"
