+++ app/src/main/java/com/example/linkybproject/auth/LoginActivity.kt
@@ -5,8 +5,7 @@ import android.content.Intent.FLAG_ACTIVITY_NEW_TASK
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.Toast
-import com.example.linkybproject.MainActivity
-import com.example.linkybproject.R
+import com.example.linkybproject.common.MainActivity
 import com.example.linkybproject.databinding.ActivityLoginBinding
 
 class LoginActivity : AppCompatActivity(), LoginView {
