+++ app/src/main/java/com/example/linkybproject/common/NetworkModule.kt
@@ -2,16 +2,14 @@ package com.example.linkybproject
 
 import com.google.gson.Gson
 import com.google.gson.GsonBuilder
-import okhttp3.Interceptor
 import okhttp3.OkHttpClient
-import okhttp3.Request
 import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 import javax.net.ssl.HostnameVerifier
 import javax.net.ssl.HttpsURLConnection
 
-const val BASE_URL = "http://192.168.0.3:8001"
+const val BASE_URL = "http://54.180.121.247:8001"
 
 fun getRetrofit(): Retrofit {
 
@@ -22,19 +20,16 @@ fun getRetrofit(): Retrofit {
     val client: OkHttpClient = OkHttpClient.Builder()
         .addInterceptor(interceptor)
         .hostnameVerifier(
-            HostnameVerifier { hostname, session ->
+            HostnameVerifier { _, session ->
                 val hv = HttpsURLConnection.getDefaultHostnameVerifier()
-                hv.verify("seolmunzip.shop", session)
+                hv.verify("linkyB.shop", session)
             })
         .build()
 
     val gson: Gson = GsonBuilder().setLenient().create()
 
-    val retrofit = Retrofit.Builder().baseUrl(BASE_URL)
+    return Retrofit.Builder().baseUrl(BASE_URL)
         .addConverterFactory(GsonConverterFactory.create(gson))
-//        .client(unsafeOkHttpClient().build())
         .client(client)
         .build()
-
-    return retrofit
 }
