+++ app/src/main/java/com/example/linkybproject/common/ReportBlockDialog.kt
@@ -1,4 +1,4 @@
-package com.example.linkybproject
+package com.example.linkybproject.common
 
 import android.app.Dialog
 import android.graphics.Color
