+++ app/src/main/java/com/example/linkybproject/myprofile/MyProfile1Fragment.kt
@@ -7,7 +7,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.example.linkybproject.MainActivity
+import com.example.linkybproject.common.MainActivity
 import com.example.linkybproject.databinding.FragmentMyprofile1Binding
 
 class MyProfile1Fragment : Fragment() {
